backend/requirements.txt
fastapi
uvicorn
posthog
python-dotenv
requests
aiohttp
# AI and Machine Learning Components
openai>=1.0.0
langchain>=0.1.0
langchain-openai
langchain-community
scikit-learn==1.3.0
pandas
numpy
asyncio-mqtt
redis==5.0.1
celery==5.3.4
# Pattern Recognition & Anomaly Detection
scipy
matplotlib
seaborn==0.12.2
plotly==5.17.0
# Time Series Analysis
prophet
statsmodels
# Database for agent memory
sqlalchemy
alembic
# Advanced Analytics & ML Dependencies
joblib>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0
category-encoders>=2.6.0
# Real-time Processing
kafka-python>=2.0.0
websockets>=11.0.0
# Monitoring & Logging
prometheus-client>=0.17.0
structlog>=23.1.0
# Security & Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
# HTTP & CORS
python-multipart>=0.0.6
httpx>=0.24.0